feat(theme): DS accents (brass + signals) with tier fallback in the shared theme
Consolidate the design-system identity into the pre-existing shared theme
(crate::theme, used by both the TUI and the shell) instead of a parallel
module. Accents (selection/heading/roles/tag → brass, working→running,
done→ok, error→critical, warning) resolve from the DS palette with a
truecolor→256→16 tier fallback (detected once via theme::init). muted and text
stay NAMED (adaptive) so they remain legible on light AND dark terminals — the
audit's deliberate principle. Views/shell keep the same theme::*() API.
